Rename csvparser to timestampcsvparser
timestampcsvparser is a more accurate name than csvparser because it
indicates that the parser is not generic to all CSVs.
